minor changes for the merge
