+++ core/src/main/java/net/akami/yggdrasil/spell/test/EarthTowerSpell.java
@@ -1,4 +1,4 @@
-package net.akami.yggdrasil.spell;
+package net.akami.yggdrasil.spell.test;
 
 import net.akami.yggdrasil.api.spell.Spell;
 import net.akami.yggdrasil.api.spell.SpellLauncher;
@@ -7,7 +7,7 @@ import net.akami.yggdrasil.api.spell.SpellTier;
 import java.util.Arrays;
 import java.util.List;
 
-public class WindOfFireSpell implements Spell {
+public class EarthTowerSpell implements Spell {
 
     private List<SpellTier> tiers;
 
@@ -16,11 +16,11 @@ public class WindOfFireSpell implements Spell {
         return null;
     }
 
-    public WindOfFireSpell() {
+    public EarthTowerSpell() {
         this.tiers = Arrays.asList(
-                new WindOfFireTestTier(),
-                new WindOfFireTestTier(),
-                new WindOfFireTestTier());
+                new EarthTowerTestTier(),
+                new EarthTowerTestTier(),
+                new EarthTowerTestTier());
     }
 
     @Override
