+++ core/src/main/java/net/akami/yggdrasil/spell/test/GravitySpell.java
@@ -1,4 +1,5 @@
-package net.akami.yggdrasil.spell;
+package net.akami.yggdrasil.spell.test;
+
 
 import net.akami.yggdrasil.api.spell.Spell;
 import net.akami.yggdrasil.api.spell.SpellLauncher;
@@ -7,24 +8,24 @@ import net.akami.yggdrasil.api.spell.SpellTier;
 import java.util.Arrays;
 import java.util.List;
 
-public class EarthTowerSpell implements Spell {
+public class GravitySpell implements Spell {
 
     private List<SpellTier> tiers;
 
-    @Override
-    public SpellLauncher getLauncher() {
-        return null;
-    }
-
-    public EarthTowerSpell() {
+    public GravitySpell() {
         this.tiers = Arrays.asList(
-                new EarthTowerTestTier(),
-                new EarthTowerTestTier(),
-                new EarthTowerTestTier());
+                new GravityTestTier(),
+                new GravityTestTier(),
+                new GravityTestTier());
     }
 
     @Override
     public List<SpellTier> getTiers() {
         return tiers;
+    }
+
+    @Override
+    public SpellLauncher getLauncher() {
+        return null;
     }
 }
