+++ core/src/main/java/net/akami/yggdrasil/spell/test/WindOfFireSpell.java
@@ -1,5 +1,4 @@
-package net.akami.yggdrasil.spell;
-
+package net.akami.yggdrasil.spell.test;
 
 import net.akami.yggdrasil.api.spell.Spell;
 import net.akami.yggdrasil.api.spell.SpellLauncher;
@@ -8,24 +7,24 @@ import net.akami.yggdrasil.api.spell.SpellTier;
 import java.util.Arrays;
 import java.util.List;
 
-public class GravitySpell implements Spell {
+public class WindOfFireSpell implements Spell {
 
     private List<SpellTier> tiers;
 
-    public GravitySpell() {
+    @Override
+    public SpellLauncher getLauncher() {
+        return null;
+    }
+
+    public WindOfFireSpell() {
         this.tiers = Arrays.asList(
-                new GravityTestTier(),
-                new GravityTestTier(),
-                new GravityTestTier());
+                new WindOfFireTestTier(),
+                new WindOfFireTestTier(),
+                new WindOfFireTestTier());
     }
 
     @Override
     public List<SpellTier> getTiers() {
         return tiers;
-    }
-
-    @Override
-    public SpellLauncher getLauncher() {
-        return null;
     }
 }
